requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-filter==22.1
django-searchable-select==1.5.0
frozenlist==1.3.3
idna==3.4
lxml==4.9.2
magic-filter==1.0.9
multidict==6.0.4
novaposhta-api-client @ git+https://github.com/semolex/novaposhta-api-client@8dd77ed29e6bb0ef3c7b931f79e33e3d99f55117
pyTelegramBotAPI==4.10.0
pytz==2022.7.1
requests==2.28.2
soupsieve==2.4
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.15
yarl==1.8.2
